fix(progressbar): fiz progressbar issue
