+++ components/shared/hexToRgb.js
@@ -1,0 +1,11 @@
+const hexToRgb = hex => {
+  const result = /^#?([a-f\d]{2})([a-f\d]{2})([a-f\d]{2})$/i.exec(hex);
+  return result
+    ? `rgb(${parseInt(result[1], 16)}, ${parseInt(result[2], 16)}, ${parseInt(
+        result[3],
+        16,
+      )})`
+    : 'rgba(255, 255, 255)';
+};
+
+export default hexToRgb;
