+++ components/shared/index.js
@@ -8,6 +8,7 @@ import BREAKPOINTS from './breakpoints';
 import uniqId from './uniqId';
 import * as INPUT_STYLE from './inputStyle';
 import hexToRgba from './hexToRgba';
+import hexToRgb from './hexToRgb';
 import shadow from './shadow';
 import HiddenInput from './HiddenInput';
 
@@ -22,6 +23,7 @@ export {
   INPUT_STYLE,
   uniqId,
   hexToRgba,
+  hexToRgb,
   shadow,
   HiddenInput,
 };
