--- shared/helpers.js
@@ -1,21 +1,0 @@
-const hexToRgb = hex => {
-  const result = /^#?([a-f\d]{2})([a-f\d]{2})([a-f\d]{2})$/i.exec(hex);
-  return result
-    ? `rgb(${parseInt(result[1], 16)}, ${parseInt(result[2], 16)}, ${parseInt(
-        result[3],
-        16,
-      )})`
-    : 'rgba(255, 255, 255)';
-};
-
-const hexToRgba = (hex, opacity) => {
-  const result = /^#?([a-f\d]{2})([a-f\d]{2})([a-f\d]{2})$/i.exec(hex);
-  return result
-    ? `rgba(${parseInt(result[1], 16)}, ${parseInt(result[2], 16)}, ${parseInt(
-        result[3],
-        16,
-      )}, ${opacity})`
-    : 'rgba(255, 255, 255, 1)';
-};
-
-export { hexToRgb, hexToRgba };
